more (un)signed fixes.

+++ gifs_src/gifs/gifs.cpp
@@ -3,7 +3,7 @@
 #include "properties.hpp"
 
 
-void create_qm_interface(int nqm, int* qm_atomids)
+void create_qm_interface(size_t nqm, int* qm_atomids)
 {
     if (qm == nullptr) {
         std::vector<int> qmids(qm_atomids, qm_atomids+nqm);
@@ -12,10 +12,10 @@ void create_qm_interface(int nqm, int* qm_atomids)
 }
 
 
-float gifs_get_forces(float* qm_crd, int nmm, float* mm_crd, float* mm_chg, float* f, float* fshift)
+float gifs_get_forces(float* qm_crd, size_t nmm, float* mm_crd, float* mm_chg, float* f, float* fshift)
 {
     //
-    int nqm = qm->get_nqm();
+    size_t nqm = qm->get_nqm();
     qm->update(qm_crd, nmm, mm_crd, mm_chg);
     // 
     std::vector<double> g_qm(nqm*3), g_mm(3*nmm);
